+++ src/api.ts
@@ -1,16 +1,61 @@
 import { Api } from "coder/site/src/api/api"
+import { ProvisionerJobLog, Workspace } from "coder/site/src/api/typesGenerated"
 import fs from "fs/promises"
 import * as os from "os"
 import { ProxyAgent } from "proxy-agent"
-import { getProxyForUrl } from "proxy-from-env"
 import * as vscode from "vscode"
+import * as ws from "ws"
+import { errToStr } from "./api-helper"
 import { CertificateError } from "./error"
+import { getProxyForUrl } from "./proxy"
 import { Storage } from "./storage"
 
 // expandPath will expand ${userHome} in the input string.
-const expandPath = (input: string): string => {
+function expandPath(input: string): string {
   const userHome = os.homedir()
   return input.replace(/\${userHome}/g, userHome)
+}
+
+async function createHttpAgent(): Promise<ProxyAgent> {
+  const cfg = vscode.workspace.getConfiguration()
+  const insecure = Boolean(cfg.get("coder.insecure"))
+  const certFile = expandPath(String(cfg.get("coder.tlsCertFile") ?? "").trim())
+  const keyFile = expandPath(String(cfg.get("coder.tlsKeyFile") ?? "").trim())
+  const caFile = expandPath(String(cfg.get("coder.tlsCaFile") ?? "").trim())
+
+  return new ProxyAgent({
+    // Called each time a request is made.
+    getProxyForUrl: (url: string) => {
+      const cfg = vscode.workspace.getConfiguration()
+      return getProxyForUrl(url, cfg.get("http.proxy"), cfg.get("coder.proxyBypass"))
+    },
+    cert: certFile === "" ? undefined : await fs.readFile(certFile),
+    key: keyFile === "" ? undefined : await fs.readFile(keyFile),
+    ca: caFile === "" ? undefined : await fs.readFile(caFile),
+    // rejectUnauthorized defaults to true, so we need to explicitly set it to
+    // false if we want to allow self-signed certificates.
+    rejectUnauthorized: !insecure,
+  })
+}
+
+let agent: Promise<ProxyAgent> | undefined = undefined
+async function getHttpAgent(): Promise<ProxyAgent> {
+  if (!agent) {
+    vscode.workspace.onDidChangeConfiguration((e) => {
+      if (
+        // http.proxy and coder.proxyBypass are read each time a request is
+        // made, so no need to watch them.
+        e.affectsConfiguration("coder.insecure") ||
+        e.affectsConfiguration("coder.tlsCertFile") ||
+        e.affectsConfiguration("coder.tlsKeyFile") ||
+        e.affectsConfiguration("coder.tlsCaFile")
+      ) {
+        agent = createHttpAgent()
+      }
+    })
+    agent = createHttpAgent()
+  }
+  return agent
 }
 
 /**
@@ -31,25 +76,10 @@ export async function makeCoderSdk(baseUrl: string, token: string | undefined, s
       config.headers[key] = value
     })
 
-    const cfg = vscode.workspace.getConfiguration()
-    const insecure = Boolean(cfg.get("coder.insecure"))
-    const certFile = expandPath(String(cfg.get("coder.tlsCertFile") ?? "").trim())
-    const keyFile = expandPath(String(cfg.get("coder.tlsKeyFile") ?? "").trim())
-    const caFile = expandPath(String(cfg.get("coder.tlsCaFile") ?? "").trim())
-
     // Configure proxy and TLS.
-    const agent = new ProxyAgent({
-      // If the proxy setting exists, we always use it.  Otherwise we follow the
-      // standard environment variables (no_proxy, http_proxy, etc).
-      getProxyForUrl: (url: string) => cfg.get("http.proxy") || getProxyForUrl(url),
-      cert: certFile === "" ? undefined : await fs.readFile(certFile),
-      key: keyFile === "" ? undefined : await fs.readFile(keyFile),
-      ca: caFile === "" ? undefined : await fs.readFile(caFile),
-      // rejectUnauthorized defaults to true, so we need to explicitly set it to
-      // false if we want to allow self-signed certificates.
-      rejectUnauthorized: !insecure,
-    })
-
+    // Note that by default VS Code overrides the agent.  To prevent this, set
+    // `http.proxySupport` to `on` or `off`.
+    const agent = await getHttpAgent()
     config.httpsAgent = agent
     config.httpAgent = agent
 
@@ -65,4 +95,89 @@ export async function makeCoderSdk(baseUrl: string, token: string | undefined, s
   )
 
   return restClient
+}
+
+/**
+ * Start or update a workspace and return the updated workspace.
+ */
+export async function startWorkspace(restClient: Api, workspace: Workspace): Promise<Workspace> {
+  // If the workspace requires the latest active template version, we should attempt
+  // to update that here.
+  // TODO: If param set changes, what do we do??
+  const versionID = workspace.template_require_active_version
+    ? // Use the latest template version
+      workspace.template_active_version_id
+    : // Default to not updating the workspace if not required.
+      workspace.latest_build.template_version_id
+  const latestBuild = await restClient.startWorkspace(workspace.id, versionID)
+  return {
+    ...workspace,
+    latest_build: latestBuild,
+  }
+}
+
+/**
+ * Wait for the latest build to finish while streaming logs to the emitter.
+ *
+ * Once completed, fetch the workspace again and return it.
+ */
+export async function waitForBuild(
+  restClient: Api,
+  writeEmitter: vscode.EventEmitter<string>,
+  workspace: Workspace,
+): Promise<Workspace> {
+  const baseUrlRaw = restClient.getAxiosInstance().defaults.baseURL
+  if (!baseUrlRaw) {
+    throw new Error("No base URL set on REST client")
+  }
+
+  // This fetches the initial bunch of logs.
+  const logs = await restClient.getWorkspaceBuildLogs(workspace.latest_build.id, new Date())
+  logs.forEach((log) => writeEmitter.fire(log.output + "\r\n"))
+
+  // This follows the logs for new activity!
+  // TODO: watchBuildLogsByBuildId exists, but it uses `location`.
+  //       Would be nice if we could use it here.
+  let path = `/api/v2/workspacebuilds/${workspace.latest_build.id}/logs?follow=true`
+  if (logs.length) {
+    path += `&after=${logs[logs.length - 1].id}`
+  }
+
+  await new Promise<void>((resolve, reject) => {
+    try {
+      const baseUrl = new URL(baseUrlRaw)
+      const proto = baseUrl.protocol === "https:" ? "wss:" : "ws:"
+      const socketUrlRaw = `${proto}//${baseUrl.host}${path}`
+      const socket = new ws.WebSocket(new URL(socketUrlRaw), {
+        headers: {
+          "Coder-Session-Token": restClient.getAxiosInstance().defaults.headers.common["Coder-Session-Token"] as
+            | string
+            | undefined,
+        },
+        followRedirects: true,
+      })
+      socket.binaryType = "nodebuffer"
+      socket.on("message", (data) => {
+        const buf = data as Buffer
+        const log = JSON.parse(buf.toString()) as ProvisionerJobLog
+        writeEmitter.fire(log.output + "\r\n")
+      })
+      socket.on("error", (error) => {
+        reject(
+          new Error(`Failed to watch workspace build using ${socketUrlRaw}: ${errToStr(error, "no further details")}`),
+        )
+      })
+      socket.on("close", () => {
+        resolve()
+      })
+    } catch (error) {
+      // If this errors, it is probably a malformed URL.
+      reject(new Error(`Failed to watch workspace build on ${baseUrlRaw}: ${errToStr(error, "no further details")}`))
+    }
+  })
+
+  writeEmitter.fire("Build complete\r\n")
+  const updatedWorkspace = await restClient.getWorkspace(workspace.id)
+  writeEmitter.fire(`Workspace is now ${updatedWorkspace.latest_build.status}\r\n`)
+  return updatedWorkspace
 }